+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Only print errors in red if the output terminal supports ANSI colour codes.
   [#58](https://github.com/kylef/Commander/pull/58)
 
+- `ArgumentParser.isEmpty` will now return empty for empty arguments.
+
 ## 0.8.0
 
 ### Enhancements